pyproject.toml
[project]
name = "pyfatsecret-chocotonic"
version = "0.1.0"
description = "Python wrapper for the Fatsecret API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black",
    "bs4",
    "isort",
    "lxml",
    "pytest-cov",
    "pytest",
    "python-dotenv",
    "requests-oauthlib",
    "requests",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
    "integration: mark a test as an integration test"
]